+++ Clients/Client.Main/ViewModels/CompraResultadoBusquedaViewModel.cs
@@ -68,7 +68,7 @@ namespace Client.Main.ViewModels
                 short a;
                 if (Int16.TryParse(value, out a))
                 {
-                    if (Int16.Parse(value) < DateTime.Now.Day)
+                    if (Int32.Parse(value) < DateTime.Now.Day)
                     {
                         _dia = value;
                         NotifyOfPropertyChange(() => Dia);
@@ -87,7 +87,7 @@ namespace Client.Main.ViewModels
                 short a;
                 if (Int16.TryParse(value, out a))
                 {
-                    if (Int16.Parse(value) <= DateTime.Now.Month)
+                    if (Int32.Parse(value) <= DateTime.Now.Month)
                     {
                         _mes = value;
                         NotifyOfPropertyChange(() => Mes);
@@ -108,7 +108,7 @@ namespace Client.Main.ViewModels
                 short a;
                 if (Int16.TryParse(value, out a))
                 {
-                    if (Int16.Parse(value) < DateTime.Now.Year)
+                    if (Int32.Parse(value) < DateTime.Now.Year)
                     {
                         _año = value;
                         NotifyOfPropertyChange(() => Año);
